% Generated by roxygen2 (4.0.1): do not edit by hand
\name{errorGen}
\alias{errorGen}
\title{Generate Error statistics}
\usage{
errorGen(finalraster, measured.spdf, measured.data, plot = FALSE,
  title = "")
}
\arguments{
  \item{finalraster}{RasterLayer object}

  \item{measured.spdf}{SpatialPointsDataFrame}

  \item{measured.data}{data.frame}

  \item{plot}{logical. Plot comparison?}

  \item{title}{Plot labels}
}
\value{
List of error statistics
}
\description{
Generate Error statistics
}
\author{
Joseph Stachelek
}

