% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipdwInterp.R
\name{rm_na_pointslayers}
\alias{rm_na_pointslayers}
\title{Remove NA points features and drop corresponding raster stack layers}
\usage{
rm_na_pointslayers(param_name, sf_ob, rstack)
}
\arguments{
\item{param_name}{character name of data column}

\item{sf_ob}{sf object with point geometries}

\item{rstack}{RasterStack or RasterBrick}
}
\description{
Remove NA points features and drop corresponding raster stack layers
}
