% Generated by roxygen2 (4.0.1): do not edit by hand
\name{costrasterGen}
\alias{costrasterGen}
\title{Cost Raster generation}
\usage{
costrasterGen(xymat, pols, projstr)
}
\arguments{
  \item{xymat}{Matrix of coordinates}

  \item{pols}{SpatialPolygons object}

  \item{projstr}{proj4 string defining the inherent
  projection}
}
\value{
RasterLayer
}
\description{
Cost Raster generation
}
\author{
Joseph Stachelek
}

