% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ipdw}
\alias{ipdw}
\title{Inverse Path Distance Weighting}
\usage{
ipdw(spdf, costras, range, paramlist, overlapped = FALSE,
  yearmon = "default", removefile = TRUE, step = 16)
}
\arguments{
  \item{spdf}{SpatialPointsDataFrame object}

  \item{costras}{RasterLayer. Cost raster}

  \item{range}{numeric. Range of interpolation
  neighborhood}

  \item{paramlist}{character. String representing parameter
  names}

  \item{overlapped}{logical. Default is FALSE, specify TRUE
  if some points lie on top of barriers}

  \item{yearmon}{character. String specifying the name of
  the spdf}

  \item{removefile}{logical. Remove files after
  processing?}

  \item{step}{numeric. Number of sub loops to manage memory
  during raster processing.}
}
\value{
RasterStack
}
\description{
Inverse Path Distance Weighting
}
\details{
This is a high level function that interpolates a
SpatialPointsDataFrame object in a single pass.
}
\examples{
#see vignette
}
\author{
Joseph Stachelek
}

