% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ipdwInterp}
\alias{ipdwInterp}
\title{Inverse Distance Weighting with custom distances}
\usage{
ipdwInterp(spdf, rstack, paramlist, overlapped = FALSE, yearmon = "default",
  removefile = TRUE)
}
\arguments{
  \item{spdf}{SpatialPointsDataFrame object}

  \item{rstack}{RasterStack of path distances}

  \item{paramlist}{chacter. String representing parameter
  names}

  \item{overlapped}{logical. Default is FALSE, specify TRUE
  if some points lie on top of barriers}

  \item{yearmon}{character. String specifying the name of
  the spdf}

  \item{removefile}{logical. Remove files after
  processing?}
}
\value{
RasterLayer
}
\description{
Inverse Distance Weighting with custom distances
}
\author{
Joseph Stachelek
}

