% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pathdistGen}
\alias{pathdistGen}
\title{Path Distance generation}
\usage{
pathdistGen(spdf, costras, range, step = 16, yearmon = "default")
}
\arguments{
  \item{spdf}{SpatialPointsDataFrame object}

  \item{costras}{RasterLayer cost raster}

  \item{range}{numeric. Range of interpolation
  neighborhood}

  \item{step}{numeric. Number of sub loops to manage memory
  during raster processing.}

  \item{yearmon}{character. String specifying the name of
  the spdf}
}
\value{
RasterStack object of path distances
}
\description{
Path Distance generation
}
\author{
Joseph Stachelek
}

